Refactor: separating register and edit botton from admin page
